feat: add id check

+++ backend/controllers/product.controller.js
@@ -56,11 +56,16 @@ const editProduct = async (req, res) => {
 const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  if (!mongoose.Types.ObjectId.isValid(id))
+    return res
+      .status(400)
+      .json({ success: false, message: "Product not found" });
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: "Product deleted" });
   } catch (error) {
-    res.status(404).json({ success: false, message: "Product not found" });
+    res.status(500).json({ success: false, message: "Server error" });
     console.error(`Error in deletion: ${error.message}`);
   }
 };
